+++ NEWS.md
@@ -1,3 +1,3 @@
 # rstudiothemes (development version)
 
--   First stable version.
+-   In active development.